src/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prose-arc-backend"
version = "0.1.0"
description = "Prose Arc — novel writing SaaS backend"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "itsdangerous>=2.2.0",   # required by Starlette SessionMiddleware

    # Database
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Validation & settings
    "pydantic>=2.9.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Auth
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "authlib>=1.3.2",
    "httpx>=0.27.0",  # async HTTP client (OAuth + tests)

    # Async tasks
    "celery[redis]>=5.4.0",
    "kombu>=5.3.0",
    "redis[asyncio]>=5.0.0",

    # Object storage (MinIO via S3, GCS)
    "boto3>=1.35.0",

    # Logging
    "structlog>=24.4.0",

    # Utilities
    "python-multipart>=0.0.12",  # form data support for FastAPI
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "factory-boy>=3.3.0",
    "anyio>=4.6.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in default arguments (FastAPI pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ---------------------------------------------------------------------------
# mypy — strict type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires running external services (postgres, redis, minio)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "migrations/*", "scripts/*"]
